% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesar_helpers.R
\name{cesar_gradient_n_pal}
\alias{cesar_gradient_n_pal}
\title{Cesar gradient palette}
\usage{
cesar_gradient_n_pal(
  palette = "cesar_green_c",
  alpha = 1,
  reverse = T,
  na.color = NA
)
}
\arguments{
\item{palette}{Choose from 'cesar_palettes' list}

\item{alpha}{sets transparency of each color}

\item{reverse}{if true, reverses the order of palette}

\item{na.color}{color for NA values}
}
\description{
Helper function to incorporate Cesar colors into scale fills
}
