% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesar_helpers.R
\name{cesar_pal}
\alias{cesar_pal}
\title{Cesar palette with ramped colors}
\usage{
cesar_pal(palette = "cesar", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'cesar_palettes' list}

\item{alpha}{sets transparency of each color}

\item{reverse}{if true, reverses the order of palette}
}
\description{
Helper function to incorporate Cesar colors into scale fills
}
