% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesar_colours.R
\docType{data}
\name{cesar_palettes}
\alias{cesar_palettes}
\title{List of Cesar Colour Palettes}
\format{
An object of class \code{list} of length 13.
}
\usage{
cesar_palettes
}
\description{
Sets of colour palettes for use in Cesar Australia data visualisations.
Colours are stored as a named list. Use palette name or position to call in
colours.

Colour sets are stored as the name of the first colour:
Calling \code{cesar_palettes$cesar} will return the seven color palette where Cesar green is the base.
Calling \code{cesar_palettes$cesar_green_c} will return an interpolated ramp of 10 colors for Cesar green.
}
\examples{
\dontrun{

# List names of palettes available
names(cesar_palettes)

show_col(cesar_palettes$cesar)
show_col(cesar_palettes$gold_teal_d)
show_col(cesar_palettes$warm_grey_c)
}
}
\keyword{datasets}
