% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesar_helpers.R
\name{cesar_rampr}
\alias{cesar_rampr}
\title{Palette interpolation}
\usage{
cesar_rampr(n, pal_name = "cesar", alpha = 1, reverse = FALSE)
}
\arguments{
\item{n}{number of interpolation points}

\item{pal_name}{name of the palette}

\item{alpha}{sets transparency of each color}

\item{reverse}{if true, reverses order}
}
\description{
Generate a palette by linear interpolation.
If a discrete palette is selected, colors are recycled to the length of the desired vector.
}
\details{
Chooses palette to use with ggplot2
}
\examples{
\dontrun{
#View different interpolations of palettes
cesar_rampr("cesar_green_c") \%>\%  scales::show_col()
cesar_rampr("cesar_green_c", n = 10) \%>\% scales::show_col()
cesar_rampr("cesar_green_c", n = 100)
}
}
