% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesar_save.R
\name{cesar_save}
\alias{cesar_save}
\title{Save Plot}
\usage{
cesar_save(
  filename,
  plot = last_plot(),
  path = NULL,
  preset = NULL,
  logo = FALSE,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk, png unless otherwise specified}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{path}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.}

\item{preset}{Passes pre-determined values to the arguments \code{scale}, \code{width}, \code{height} and \code{dpi} based on the type of output the visualisation is required for. Accepts one of \code{c("twitter", "web", "linkedin", "facebook", "print")} and overwrites user-provided values. If \code{NULL} (default) inherits \code{ggplot2::ggsave} default values.}

\item{logo}{if \code{TRUE} add Cesar logo at top right. Defaults to \code{FALSE}}

\item{...}{additional arguments to that pass to ggsave}
}
\value{
exports a file with Cesar dimensions
}
\description{
Wrapper around \code{ggplot2::ggsave} with Cesar standards
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + theme_cesar()
#default save
cesar_save("out/sepal.png")
#defaults to png if not specified (.pdf, .svg, .jpeg)
cesar_save("out/sepal")
#change the preset
cesar_save("out/sepal.jpeg", preset = "twitter")  }
}
\concept{saving}
