% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\docType{data}
\name{color_gridline}
\alias{color_gridline}
\title{Gridline color default}
\format{
An object of class \code{character} of length 1.
}
\usage{
color_gridline
}
\description{
sets gridline colors
}
\keyword{internal}
