% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\docType{data}
\name{color_strip_background}
\alias{color_strip_background}
\title{Strip background color default}
\format{
An object of class \code{character} of length 1.
}
\usage{
color_strip_background
}
\description{
sets strip background colors
}
\keyword{internal}
