% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\docType{data}
\name{color_strip_text}
\alias{color_strip_text}
\title{Strip text color default}
\format{
An object of class \code{character} of length 1.
}
\usage{
color_strip_text
}
\description{
sets strip text colors
}
\keyword{internal}
