% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\docType{data}
\name{color_subtitle}
\alias{color_subtitle}
\title{Subtitle colour default}
\format{
An object of class \code{character} of length 1.
}
\usage{
color_subtitle
}
\description{
sets color caption
}
\keyword{internal}
