% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesar_colours.R
\docType{data}
\name{dark_grey}
\alias{dark_grey}
\title{dark grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
dark_grey
}
\description{
Dark grey
}
\seealso{
Other Cesar colors: 
\code{\link{adobe_avenue}},
\code{\link{ancient_lavastone}},
\code{\link{cesar_green}},
\code{\link{defense_matrix}},
\code{\link{galliano}},
\code{\link{gold_bullion}},
\code{\link{hobgoblin}},
\code{\link{ice_cold}},
\code{\link{key_lime}},
\code{\link{light_grey}},
\code{\link{mid_grey}},
\code{\link{oxford_blue}},
\code{\link{rufous}},
\code{\link{warm_grey}}
}
\concept{Cesar colors}
\keyword{internal}
