% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_elements.R
\name{drop_legend}
\alias{drop_legend}
\alias{drop_grid}
\alias{drop_labels}
\alias{drop_titles}
\title{Drop static elements from plots}
\usage{
drop_legend(axis = NULL)

drop_grid(axis = NULL)

drop_labels(axis = NULL)

drop_titles(axis = NULL)
}
\arguments{
\item{axis}{either "x", "y" or "both"}
}
\description{
Drop a static theme element from a ggplot object
}
\examples{
library(ggplot2)
p <- ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
 geom_point() + theme_cesar()

p + drop_legend()

p + drop_grid("x")

p + drop_labels("y")

p + drop_titles("both")

}
