% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cesar.R
\name{theme_cesar}
\alias{theme_cesar}
\title{Create Standard Cesar Plots}
\usage{
theme_cesar(
  font_title,
  font_subtitle,
  font_plot,
  font_caption,
  facet_space = 2,
  text_scale = 1,
  ...
)
}
\arguments{
\item{font_title}{font family for titles, defaults to Inter/Arial}

\item{font_subtitle}{font family for subtitles, defaults to Inter/Arial}

\item{font_plot}{font family for the plot space, defaults to Inter/Arial}

\item{font_caption}{font family for captions, defaults to Inter/Arial}

\item{facet_space}{controls how far apart facets are from each other}

\item{text_scale}{scalar that will grow/shrink all text defined within}

\item{...}{pass any parameters from theme that are not already defined within}
}
\description{
This function converts your ggplot object to the Cesar theme. The default font is Inter
}
\examples{
library(systemfonts)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
 geom_point() + theme_cesar()

}
